fix: configure private recycle repository using %U
Set up a private recycle repository for each user. This is necessary
because the vfs_recycle module does not preserve ACL inheritance when
using the `recycle:keeptree = yes` option. As a result, the contents of
the recycle bin are readable only by the user who deleted them and the
Administrator.